@@ -9,15 +9,15 @@
     "preview": "vite preview"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^3.0.2",
+    "@sveltejs/vite-plugin-svelte": "^5.0.3",
     "@vitejs/plugin-basic-ssl": "^1.1.0",
-    "svelte": "^4.2.12",
-    "vite": "^5.1.6"
+    "svelte": "^5.0.0",
+    "vite": "^6.3.2"
   },
   "dependencies": {
-    "@editorjs/editorjs": "^2.29.1",
     "@editorjs/checklist": "^1.6.0",
     "@editorjs/code": "^2.9.0",
+    "@editorjs/editorjs": "^2.29.1",
     "@editorjs/header": "^2.8.1",
     "@editorjs/image": "^2.9.0",
     "@editorjs/inline-code": "^1.5.0",
